+++ src/main/java/com/project/carPoor/repository/MemberRepository.java
@@ -3,8 +3,10 @@ package com.project.carPoor.repository;
 
 import com.project.carPoor.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
+
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
